Bootstrap pip for Python 3.8 to fix incompat pip version
On Python 3.8, the Travis-provided pip is broken since it uses a
deprecated `platform.linux_distribution()`. To work around this,
we attempt to install pip using the get-pip.py script. The script
contains the latest copy of pip, which should be compatible
with Python 3.8 and allow us to install pip itself into our
environment, and thus let us install the packages we need to
run the tests successfully.