Fix virtualenv version to allow Python 3.8 compat
Actual problem was with the virtualenv version - <14.0 does not
contain the latest pip, hence the breakage. <14.0 doesn't seem
to make sense here, since we're not supporting Python 3.2 anymore.
In the rare case that Ubuntu Xenial's version is used (it shouldn't,
Travis will usually override that), it also ships with Python 3.5,
so this should be OK.

Removed older logic of using get-pip.py - no longer needed.