Remove version locks on development pkg install
Version locks were previously added due to Python 3.4
requirements, but have been removed since Python 3.4 is EOL'd,
and the various verison requirements no longer make sense.

The version lock is still maintained for the initial install
as to provide a solid base to build on for the mini venv install.