borg repository might be local or remote, fixes #47
the repo location is just a str we pass to the borg commandline,
nothing special (like a Path object).

it might be a local path, but it also can be e.g. ssh://... .

because we change the cwd when importing backups,
a local borg repo must be given via an ABSOLUTE path.